requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pydantic-ai==0.0.14
pydantic==2.5.0
google-generativeai==0.3.2